Conflicts:
	MonoGame.Framework/Android/AndroidGameActivity.cs

+++ MonoGame.Framework.Content.Pipeline/Serialization/Intermediate/ByteSerializer.cs
@@ -8,19 +8,19 @@ using System.Xml;
 namespace Microsoft.Xna.Framework.Content.Pipeline.Serialization.Intermediate
 {
     [ContentTypeSerializer]
-    class FloatSerializer : ElementSerializer<float>
+    class ByteSerializer : ElementSerializer<byte>
     {
-        public FloatSerializer() :
-            base("float", 1)
+        public ByteSerializer() :
+            base("byte", 1)
         {
         }
 
-        protected internal override float Deserialize(string[] inputs, ref int index)
+        protected internal override byte Deserialize(string[] inputs, ref int index)
         {
-            return XmlConvert.ToSingle(inputs[index++]);
+            return XmlConvert.ToByte(inputs[index++]);
         }
 
-        protected internal override void Serialize(float value, List<string> results)
+        protected internal override void Serialize(byte value, List<string> results)
         {
             results.Add(XmlConvert.ToString(value));
         }
