+++ MonoGame.Framework.Content.Pipeline/Serialization/Intermediate/ShortSerializer.cs
@@ -8,19 +8,19 @@ using System.Xml;
 namespace Microsoft.Xna.Framework.Content.Pipeline.Serialization.Intermediate
 {
     [ContentTypeSerializer]
-    class FloatSerializer : ElementSerializer<float>
+    class ShortSerializer : ElementSerializer<short>
     {
-        public FloatSerializer() :
-            base("float", 1)
+        public ShortSerializer() :
+            base("short", 1)
         {
         }
 
-        protected internal override float Deserialize(string[] inputs, ref int index)
+        protected internal override short Deserialize(string[] inputs, ref int index)
         {
-            return XmlConvert.ToSingle(inputs[index++]);
+            return XmlConvert.ToInt16(inputs[index++]);
         }
 
-        protected internal override void Serialize(float value, List<string> results)
+        protected internal override void Serialize(short value, List<string> results)
         {
             results.Add(XmlConvert.ToString(value));
         }
